new header component and css moved from Home component and Home css
